Move batch chat completion logic into OpenAIServingChatBatch subclass
